[BE] TEST: sequelize mock 제거, test seed 생성
- sequelize-mock 번거로워서 사용하지 않기로 결정
- test seed 설정
- test 전에 seed 실행하도록 구현
