+++ server/test/label.api.test.js
@@ -1,11 +1,16 @@
 require('module-alias/register');
 const request = require('supertest');
 const app = require('@root/app');
+const seeder = require('@test/test-seed');
 
-jest.mock('@models', () => {
-  // 추후 global setting 으로 한번에 설정하는 방안 생각
-  const connection = require('@test/sequelize-mock');
-  return connection;
+beforeAll(async done => {
+  await seeder.up();
+  done();
+});
+
+afterAll(async done => {
+  await seeder.down();
+  done();
 });
 
 const SUCCESS_CODE = 200;
